The manual peak detector UI has been refactored. Improvements needed.
It's not perfect though, but it has been refactored to make it easier to
improve and maintain it.
